+++ pom.xml
@@ -1,7 +1,7 @@
 <project>
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.toby.kinectvideomask</groupId>
-  <artifactId>kinectvideomask</artifactId>
+  <groupId>org.toby.blindeye</groupId>
+  <artifactId>blindeye</artifactId>
   <version>1</version>
   <build>
     <plugins>
