+++ src/org/toby/blindeye/interfaces/SoundsInterface.java
@@ -1,0 +1,5 @@
+package org.toby.blindeye.interfaces;
+
+public interface SoundsInterface {
+  void playSound();
+}
