--- src/org/toby/kinectvideomask/interfaces/LoadersInterface.java
@@ -1,8 +1,0 @@
-package org.toby.kinectvideomask.interfaces;
-
-import KinectPV2.KinectPV2;
-import processing.core.PImage;
-
-public interface LoadersInterface {
-  PImage execute(PImage liveVideo, PImage body, PImage savedBackground, KinectPV2 kinect);
-}
