--- src/org/toby/kinectvideomask/interfaces/SoundsInterface.java
@@ -1,5 +1,0 @@
-package org.toby.kinectvideomask.interfaces;
-
-public interface SoundsInterface {
-  void playSound();
-}
